Allow plugins to list permissions
