+++ src/Powercord/plugins/pc-pluginManager/scss/_plugin.scss
@@ -13,15 +13,14 @@
     }
   }
 
-  &-container {
+  &-container, &-permissions {
     display: flex;
     justify-content: space-between;
     flex-wrap: wrap;
 
     > div {
       width: 33%;
-      margin-top: 10px;
-      margin-bottom: 10px;
+      margin: 10px 0;
       display: flex;
       align-items: center;
 
@@ -43,6 +42,22 @@
 
       &.description {
         width: 100%;
+      }
+    }
+  }
+
+  &-permissions {
+    padding-top: 10px;
+    margin-top: 10px;
+    flex-wrap: nowrap;
+    flex-direction: column;
+
+    .item {
+      width: 100%;
+      margin: 5px 0;
+
+      + .item {
+        margin-left: 25px;
       }
     }
   }
@@ -73,7 +88,7 @@
   background: rgba(32, 34, 37, .6);
   border: 1px #202225 solid;
 
-  &-footer {
+  &-permissions, &-footer {
     border-top: 1px rgba(114, 118, 125, .3) solid;
   }
 }
@@ -82,7 +97,7 @@
   background: #f8f9f9;
   border: 1px #dcddde solid;
 
-  &-footer {
+  &-permissions, &-footer {
     border-top: 1px rgba(185, 187, 190, .3) solid;
   }
